refactor(consumer): retrofit ConsumerBuilder to bon
Move the JoinGroup/SyncGroup/Heartbeat build logic from the hand-rolled
ConsumerBuilder into Consumer::start annotated with #[bon::bon] /
#[builder(start_fn = builder, finish_fn = build)], matching the pattern
established in Task 2 for Client::builder. Drop ConsumerBuilder from the
public API; keep AutoOffsetReset and the codec helpers in builder.rs.
Update integration tests to use the new Consumer::builder() call shape.
