+++ crates/client-consumer/Cargo.toml
@@ -14,6 +14,7 @@ workspace = true
 default = []
 
 [dependencies]
+bon = { workspace = true }
 crabka-protocol = { version = "0.1", path = "../protocol", default-features = false }
 crabka-client-core = { version = "0.1", path = "../client-core" }
 bytes = { workspace = true }
